requirements.txt
absl-py==1.1.0
attrs==21.4.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cycler==0.11.0
fonttools==4.34.4
kiwisolver==1.4.3
matplotlib==3.5.2
mediapipe==0.8.10.1
numpy==1.23.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
wincertstore==0.2
